Remove borderRadius panel.
