+++ packages/story-editor/src/components/panels/design/sizePosition/karma/borderRadius.karma.js
@@ -49,7 +49,7 @@ describe('Border Radius', () => {
         fixture.editor.inspector.designPanel.textBox.fill
       );
 
-      const panel = fixture.editor.inspector.designPanel.borderRadius;
+      const panel = fixture.editor.inspector.designPanel.sizePosition;
       await fixture.events.click(panel.radius(), { clickCount: 3 });
       await fixture.events.keyboard.type('30');
       await fixture.events.keyboard.press('tab');
@@ -82,7 +82,7 @@ describe('Border Radius', () => {
 
     it('should allow user to add border radius for media', async () => {
       await fixture.events.click(fixture.editor.library.media.item(0));
-      const panel = fixture.editor.inspector.designPanel.borderRadius;
+      const panel = fixture.editor.inspector.designPanel.sizePosition;
 
       // Take off lock.
       await fixture.events.click(panel.lockBorderRadius);
@@ -124,7 +124,7 @@ describe('Border Radius', () => {
       fixture.editor.library.shapes.shape('Rectangle')
     );
 
-    const panel = fixture.editor.inspector.designPanel.borderRadius;
+    const panel = fixture.editor.inspector.designPanel.sizePosition;
     await fixture.events.click(panel.radius(), {
       clickCount: 3,
     });
@@ -143,6 +143,6 @@ describe('Border Radius', () => {
     await fixture.events.click(fixture.editor.library.shapesTab);
     await fixture.events.click(fixture.editor.library.shapes.shape('Circle'));
     // Verify that panel is not found.
-    expect(() => fixture.editor.inspector.designPanel.borderRadius).toThrow();
+    expect(() => fixture.editor.inspector.designPanel.sizePosition).toThrow();
   });
 });
