+++ packages/story-editor/src/elements/media/constants.js
@@ -55,7 +55,6 @@ export const resizeRules = {
 export const MEDIA_PANELS = [
   PanelTypes.FILTER,
   PanelTypes.SIZE_POSITION,
-  PanelTypes.BORDER_RADIUS,
   PanelTypes.BORDER,
   PanelTypes.ANIMATION,
   PanelTypes.LINK,
