+++ packages/story-editor/src/karma/fixture/fixture.js
@@ -200,7 +200,6 @@ export class Fixture {
 
     const panels = [
       'animation',
-      'borderRadius',
       'borderStyle',
       'captions',
       'globalStoryStyles',
